+++ src/server/sv_init.c
@@ -127,7 +127,7 @@ void SV_SetConfigstring (int index, const char *val) {
 
 void SV_GetConfigstring( int index, char *buffer, size_t bufferSize ) {
 	if ( bufferSize < 1 ) {
-		Com_Error( ERR_DROP, "SV_GetConfigstring: bufferSize == %i", bufferSize );
+		Com_Error( ERR_DROP, "SV_GetConfigstring: bufferSize == %i", (int)bufferSize );
 	}
 	if ( index < 0 || index >= MAX_CONFIGSTRINGS ) {
 		Com_Error (ERR_DROP, "SV_GetConfigstring: bad index %i", index);
@@ -155,7 +155,7 @@ void SV_SetUserinfo( int index, const char *val ) {
 
 void SV_GetUserinfo( int index, char *buffer, size_t bufferSize ) {
 	if ( bufferSize < 1 ) {
-		Com_Error( ERR_DROP, "SV_GetUserinfo: bufferSize == %i", bufferSize );
+		Com_Error( ERR_DROP, "SV_GetUserinfo: bufferSize == %i", (int)bufferSize );
 	}
 	if ( index < 0 || index >= sv_maxclients->integer ) {
 		Com_Error (ERR_DROP, "SV_GetUserinfo: bad index %i", index);
